new file for saving data and stemming
