+++ src/data/preprocessing.py
@@ -2,10 +2,11 @@ import pandas as pd
 import nltk
 from nltk.corpus import stopwords
 import string
-from nltk.tokenize import word_tokenize
-from nltk.stem import SnowballStemmer
-import download_data
-import data_from_folder
+from download_data import *
+from data_from_folder import *
+from save_data import *
+from stemming import *
+
 
 
 def dropping(dataframe) -> pd.DataFrame:
@@ -38,54 +39,6 @@ def binarization(dataframe) -> pd.DataFrame:
     dataframe.loc[dataframe['Rating'] <= 4, 'Recommended IND'] = 0
     dataframe.rename(columns={'Recommended IND': 'Top Product'}, inplace=True)
     return dataframe
-
-
-def tokenization(dataframe) -> pd.DataFrame:
-    """
-            Tokenizes the text and lower-cases it.
-
-            Args:
-                dataframe (DataFrame): Input DataFrame to be modified.
-
-            Returns:
-                dataframe: Modified DataFrame.
-    """
-    nltk.download("punkt")
-    nltk.download("stopwords")
-
-    english_sw = set(stopwords.words('english') + list(string.punctuation))
-    dataframe['Review Text'] = dataframe['Review Text'].apply(
-        lambda x: " ".join([w.lower() for w in word_tokenize(str(x)) if w.lower() not in english_sw]))
-    dataframe['Review Text'] = dataframe['Review Text'].apply(lambda text: " ".join([w for w in text if w not in "'s"]))
-
-    return dataframe
-
-def stemmed_text(dataframe) -> pd.DataFrame:
-    """
-                Stems the Review Text from the dataframe.
-
-                Args:
-                    dataframe (DataFrame): Input DataFrame to be modified.
-
-                Returns:
-                    dataframe: Modified DataFrame.
-        """
-    stem = SnowballStemmer('english')
-    dataframe['Stemmed Review Text'] = dataframe['Review Text'].apply(lambda text: " ".join([stem.stem(w) for w in text]))
-    return dataframe
-
-def new_dataset(dataframe):
-    """
-    Saves processed dataset in csv file.
-
-    Args:
-        dataframe (DataFrame): Input DataFrame to be saved.
-
-    Returns:
-        None.
-    """
-    csv_file_path = "/Users/claudialen/Dropbox/UNI/7eQuadri/TAED2/clothing-reviews/data/interim/preprocessed_data.csv"
-    dataframe.to_csv(csv_file_path, index=False)
 
 
 if __name__ == '__main__':
